feat: 컴포넌트 추상화 완료
