+++ client/src/components/ToDo/DateSelector.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import styled from 'styled-components';
+
+export default function DateSelector() {
+    return <></>;
+}
+
+const StyledWrapper = styled.div`
+    width: 100%;
+    height: 10%;
+`;
