+++ client/tsconfig.paths.json
@@ -9,7 +9,8 @@
             "@hooks/*": ["./hooks/*"],
             "@assets/*": ["./assets/*"],
             "@styles/*": ["./styles/*"],
-            "@types/*": ["./types/*"]
+            "@types/*": ["./types/*"],
+            "@utils/*": ["./utils/*"]
         }
     }
 }
